+++ src/fraiseql/__init__.py
@@ -71,7 +71,7 @@ except ImportError:
     Auth0Config = None
     Auth0Provider = None
 
-__version__ = "0.3.7"
+__version__ = "0.3.8"
 
 __all__ = [
     "ALWAYS_DATA_CONFIG",